Adding of concepts of API Versioning
In this i have shown how to add version to APIs by creating multiple
APIs or by passing version in hearders.

Basically API versioning is needed as some website get stored as cache
in some users local and if we directly chnage the Api core functionality
then that might show user some unexpected errors. So we maintain some
kind of version so that if new user comes get redirected to latest
version and old get to previous API and to inform the old user we can
notify them or change theri stored cache.